refactor: move random matrix generators to test utils
